Update API for message sending
