+++ AVOS/AVOSCloudIM/AVIMConversation.m
@@ -375,19 +375,25 @@
 - (void)sendMessage:(AVIMMessage *)message
            callback:(AVIMBooleanResultBlock)callback
 {
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
     [self sendMessage:message
               options:AVIMMessageSendOptionNone
              callback:callback];
+#pragma clang diagnostic pop
 }
 
 - (void)sendMessage:(AVIMMessage *)message
       progressBlock:(AVProgressBlock)progressBlock
            callback:(AVIMBooleanResultBlock)callback
 {
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
     [self sendMessage:message
               options:AVIMMessageSendOptionNone
         progressBlock:progressBlock
              callback:callback];
+#pragma clang diagnostic pop
 }
 
 - (void)sendMessage:(AVIMMessage *)message
@@ -405,6 +411,22 @@
       progressBlock:(AVProgressBlock)progressBlock
            callback:(AVIMBooleanResultBlock)callback
 {
+    AVIMMessageOption *option = [[AVIMMessageOption alloc] init];
+
+    if (options & AVIMMessageSendOptionTransient)
+        option.transient = YES;
+
+    if (options & AVIMMessageSendOptionRequestReceipt)
+        option.receipt = YES;
+
+    [self sendMessage:message option:option progressBlock:progressBlock callback:callback];
+}
+
+- (void)sendMessage:(AVIMMessage *)message
+             option:(AVIMMessageOption *)option
+      progressBlock:(AVProgressBlock)progressBlock
+           callback:(AVIMBooleanResultBlock)callback
+{
     message.clientId = _imClient.clientId;
     message.conversationId = _conversationId;
     if (self.imClient.status != AVIMClientStatusOpened) {
@@ -414,7 +436,6 @@
         return;
     }
     message.status = AVIMMessageStatusSending;
-
     
     if ([message isKindOfClass:[AVIMTypedMessage class]]) {
         AVIMTypedMessage *typedMessage = (AVIMTypedMessage *)message;
@@ -444,7 +465,7 @@
                         /* If uploading is success, bind file to message */
                         [self fillTypedMessage:typedMessage withFile:file];
                         [self fillTypedMessageForLocationIfNeeded:typedMessage];
-                        [self sendRealMessage:message options:options callback:callback];
+                        [self sendRealMessage:message option:option callback:callback];
                     } else {
                         message.status = AVIMMessageStatusFailed;
                         [AVIMBlockHelper callBooleanResultBlock:callback error:error];
@@ -454,14 +475,14 @@
                 /* File has already been uploaded, bind file to message */
                 [self fillTypedMessage:typedMessage withFile:file];
                 [self fillTypedMessageForLocationIfNeeded:typedMessage];
-                [self sendRealMessage:message options:options callback:callback];
+                [self sendRealMessage:message option:option callback:callback];
             }
         } else {
             [self fillTypedMessageForLocationIfNeeded:typedMessage];
-            [self sendRealMessage:message options:options callback:callback];
+            [self sendRealMessage:message option:option callback:callback];
         }
     } else {
-        [self sendRealMessage:message options:options callback:callback];
+        [self sendRealMessage:message option:option callback:callback];
     }
 }
 
@@ -548,29 +569,34 @@
     }
 }
 
-- (void)sendRealMessage:(AVIMMessage *)message options:(AVIMMessageSendOption)options callback:(AVIMBooleanResultBlock)callback {
+- (void)sendRealMessage:(AVIMMessage *)message option:(AVIMMessageOption *)option callback:(AVIMBooleanResultBlock)callback {
     dispatch_async([AVIMClient imClientQueue], ^{
-        bool transient = options & AVIMMessageSendOptionTransient;
-        bool requestReceipt = options & AVIMMessageSendOptionRequestReceipt;
+        bool transient = option.transient;
+        bool requestReceipt = option.receipt;
+
         if ([message isKindOfClass:[AVIMTypedMessage class]]) {
             AVIMTypedMessage *typedMessage = (AVIMTypedMessage *)message;
             if (!typedMessage.messageObject._lctext && !typedMessage.messageObject._lcloc && !typedMessage.messageObject._lcfile && !typedMessage.messageObject._lcattrs) {
                 [NSException raise:NSInternalInconsistencyException format:@"AVIMTypedMessage should have one of text, file, location or attributes not be nil."];
             }
         }
+
         AVIMGenericCommand *genericCommand = [[AVIMGenericCommand alloc] init];
         genericCommand.needResponse = YES;
         genericCommand.cmd = AVIMCommandType_Direct;
-        if (message.priority > 0) {
+
+        if (option.priority > 0) {
             if (self.transient) {
-                genericCommand.priority = message.priority;
+                genericCommand.priority = option.priority;
             } else {
                 AVLoggerInfo(AVLoggerDomainIM, @"Message priority has no effect in non-transient conversation.");
             }
         }
+
         AVIMDirectCommand *directCommand = [[AVIMDirectCommand alloc] init];
         [genericCommand avim_addRequiredKeyWithCommand:directCommand];
         [genericCommand avim_addRequiredKeyForDirectMessageWithMessage:message transient:NO];
+
         if (transient) {
             directCommand.transient = YES;
             genericCommand.needResponse = NO;
