+++ AVOS/AVOSCloudIM/AVIMMessageOption.m
@@ -1,0 +1,13 @@
+//
+//  AVIMMessageOption.m
+//  AVOS
+//
+//  Created by Tang Tianyong on 9/13/16.
+//  Copyright © 2016 LeanCloud Inc. All rights reserved.
+//
+
+#import "AVIMMessageOption.h"
+
+@implementation AVIMMessageOption
+
+@end
